requirements.txt
flask
flask-cors
pyjwt
pandas
scikit-learn
opencv-python
numpy
python-barcode
Pillow
